requirements.txt
aiohttp
beautifulsoup4
pika
python-dotenv
